Finally working version for ph15
Now we have an iterator that first fetches only the pk_uniqueid for all
records in the database. Then we iterate over 1000 records each. Thanks
to Edvin for this idea.
Also fix comment after volume declaration in Dockerfile: This would
result in multiple oddly-named volumes.
We now retry ldap bind forever and handle tracebacks with a sleep loop
(after logging the traceback to stdout).